+++ base/authority/src/main/java/com/kewen/framework/base/authority/utils/AuthorityConvertUtil.java
@@ -114,19 +114,19 @@ public class AuthorityConvertUtil {
             String type = split[0];
             switch (type){
                 case USER_CODE:
-                    users.add(new User(Long.valueOf(split[1]),splitDescription[1]));
+                    users.add(new User(Integer.valueOf(split[1]),splitDescription[1]));
                     break;
                 case DEPT_CODE:
-                    depts.add(new Dept(Long.valueOf(split[1]),splitDescription[1]));
+                    depts.add(new Dept(Integer.valueOf(split[1]),splitDescription[1]));
                     break;
                 case POSITION_CODE:
-                    positions.add(new Position(Long.valueOf(split[1]),splitDescription[1]));
+                    positions.add(new Position(Integer.valueOf(split[1]),splitDescription[1]));
                     break;
                 case ROLE_CODE:
-                    roles.add(new Role(Long.valueOf(split[1]),splitDescription[1]));
+                    roles.add(new Role(Integer.valueOf(split[1]),splitDescription[1]));
                     break;
                 case PERMISSION_CODE:
-                    permissions.add(new Permission(Long.valueOf(split[1]),splitDescription[1]));
+                    permissions.add(new Permission(Integer.valueOf(split[1]),splitDescription[1]));
                     break;
             }
         }
